tests/conftest.py
import pytest
from rest_framework.test import APIClient


@pytest.fixture(autouse=True)
def enable_db_access_for_all_tests(db):
    pass


@pytest.fixture
def api_client() -> APIClient:
    return APIClient()
