pyproject.toml
[tool.poetry]
name = "django-rest-batteries"
version = "0.1.0"
description = "Make DRF great again"
authors = ["Define Impossible <hi@defineimpossible.io>"]
packages = [
    { include = "rest_batteries" }
]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.6"
django = "^3.0.8"
djangorestframework = "^3.11.0"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
pytest-django = "^3.9.0"
django-debug-toolbar = "^2.2"
django-stubs = "^1.5.0"
djangorestframework-stubs = "^1.2.0"
factory-boy = "^2.12.0"
isort = "^5.2.2"
black = "^19.10b0"
brunette = "^0.1.7"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
